changes to config
