+++ network/network.go
@@ -1,5 +1,12 @@
 package network
 
+type ParsedNetwork struct {
+	Hosts   map[string]Host
+	Keys    map[string]Key
+	Network Network
+	Shapes  Shapes
+}
+
 type Network struct {
 	Domain Domain
 	P2P    P2P
